Update and rename irisData.js to data.js
